+++ src/main/java/com/example/ililbooks/domain/order/controller/OrderController.java
@@ -1,12 +1,10 @@
 package com.example.ililbooks.domain.order.controller;
 
 import com.example.ililbooks.domain.order.dto.response.OrderResponse;
-import com.example.ililbooks.domain.order.dto.response.OrdersGetResponse;
 import com.example.ililbooks.domain.order.service.OrderService;
 import com.example.ililbooks.global.dto.AuthUser;
 import com.example.ililbooks.global.dto.response.Response;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
 import org.springframework.security.access.annotation.Secured;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
@@ -31,29 +29,6 @@ public class OrderController {
         return Response.of(orderService.createOrder(authUser, pageNum, pageSize));
     }
 
-    /* 주문 단건 조회 */
-    @Secured({USER})
-    @GetMapping("/{orderId}")
-    public Response<OrderResponse> getOrder(
-            @PathVariable Long orderId,
-            @AuthenticationPrincipal AuthUser authUser,
-            @RequestParam(defaultValue = "1") int pageNum,
-            @RequestParam(defaultValue = "10") int pageSize
-    ) {
-        return Response.of(orderService.getOrder(authUser, orderId, pageNum, pageSize));
-    }
-
-    /* 주문 다건 조회 */
-    @Secured({USER})
-    @GetMapping
-    public Response<Page<OrdersGetResponse>> getOrders(
-            @AuthenticationPrincipal AuthUser authUser,
-            @RequestParam(defaultValue = "1") int pageNum,
-            @RequestParam(defaultValue = "10") int pageSize
-    ) {
-        return Response.of(orderService.getOrders(authUser, pageNum, pageSize));
-    }
-
     /* 주문 취소 */
     @Secured({USER})
     @PatchMapping("/{orderId}/cancel")
@@ -76,16 +51,5 @@ public class OrderController {
             @RequestParam(defaultValue = "10") int pageSize
     ) {
         return Response.of(orderService.updateOrderStatus(authUser, orderId, pageNum, pageSize));
-    }
-
-    /* 배송 상태 변경 (배송 대기 -> 배송 중 -> 배송완료+주문완료) */
-    @Secured({ADMIN})
-    @PatchMapping("/{orderId}/delivery")
-    public Response<OrderResponse> updateDeliveryStatus(
-            @PathVariable Long orderId,
-            @RequestParam(defaultValue = "1") int pageNum,
-            @RequestParam(defaultValue = "10") int pageSize
-    ) {
-        return Response.of(orderService.updateDeliveryStatus(orderId, pageNum, pageSize));
     }
 }
