+++ src/main/java/com/example/ililbooks/domain/order/dto/response/OrderResponse.java
@@ -7,7 +7,6 @@ import org.springframework.data.domain.Page;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
-import java.util.List;
 
 @Getter
 public class OrderResponse {
