+++ src/main/java/com/example/ililbooks/domain/order/service/OrderDeliveryService.java
@@ -1,0 +1,39 @@
+package com.example.ililbooks.domain.order.service;
+
+import com.example.ililbooks.domain.order.dto.response.OrderResponse;
+import com.example.ililbooks.domain.order.entity.Order;
+import com.example.ililbooks.domain.order.enums.DeliveryStatus;
+import com.example.ililbooks.domain.order.enums.OrderStatus;
+import com.example.ililbooks.global.exception.BadRequestException;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import static com.example.ililbooks.domain.order.enums.DeliveryStatus.DELIVERED;
+import static com.example.ililbooks.global.exception.ErrorMessage.CANNOT_DELIVER_CANCELLED_ORDER;
+
+@Service
+@RequiredArgsConstructor
+public class OrderDeliveryService {
+
+    private final OrderService orderService;
+    private final OrderGetService orderGetService;
+
+    /* 배송 상태 변경 (대기 -> 배송중 -> 배송완료) */
+    @Transactional
+    public OrderResponse updateDeliveryStatus(Long orderId, int pageNum, int pageSize) {
+        Order order = orderService.findByIdOrElseThrow(orderId);
+
+        if (order.getOrderStatus() == OrderStatus.CANCELLED) {
+            throw new BadRequestException(CANNOT_DELIVER_CANCELLED_ORDER.getMessage());
+        }
+
+        DeliveryStatus deliveryStatus = order.getDeliveryStatus().nextDeliveryStatus(order);
+        order.updateDelivery(deliveryStatus);
+
+        if (deliveryStatus == DELIVERED) {
+            order.updateOrder(OrderStatus.COMPLETE);
+        }
+        return orderGetService.getOrderResponse(order, pageNum, pageSize);
+    }
+}
